GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - main (refs/gitbutler/main)
   - artifacts/swaps/CancelSwap.ral.json
   - artifacts/swaps/ConfirmSwap.ral.json
   - contracts/swaps/swap.ral
   - artifacts/swaps/Swap.ral.json
   - artifacts/ts/Swap.ts
   - .project.json
   - scripts/2_swap_template.ts

Your previous branch was: refs/heads/main

The sha for that commit was: 6653d100a2d92135026c0676f82ccc57918345ee

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch